Move RetryJobPhase to its own module for Fast Refresh compliance.
Keeps retry-job-dialog.tsx exporting only the component per React Doctor only-export-components.
