+++ apps/web/src/components/retry-job-phase.ts
@@ -1,0 +1,7 @@
+export const RetryJobPhase = {
+  RETRYING: 'retrying',
+  SUCCESS: 'success',
+  ERROR: 'error',
+} as const
+
+export type RetryJobPhase = (typeof RetryJobPhase)[keyof typeof RetryJobPhase]
